+++ src/lib/cldr/locale-data.ts
@@ -7,88 +7,155 @@
 import enNumbers from 'cldr-numbers-full/main/en/numbers.json'
 import enCurrencies from 'cldr-numbers-full/main/en/currencies.json'
 import enDates from 'cldr-dates-full/main/en/ca-gregorian.json'
-import enLocaleNames from 'cldr-localenames-full/main/en/languages.json'
+import enLanguages from 'cldr-localenames-full/main/en/languages.json'
+import enTerritories from 'cldr-localenames-full/main/en/territories.json'
+import enScripts from 'cldr-localenames-full/main/en/scripts.json'
+import enVariants from 'cldr-localenames-full/main/en/variants.json'
 
 // German
 import deNumbers from 'cldr-numbers-full/main/de/numbers.json'
 import deCurrencies from 'cldr-numbers-full/main/de/currencies.json'
 import deDates from 'cldr-dates-full/main/de/ca-gregorian.json'
-import deLocaleNames from 'cldr-localenames-full/main/de/languages.json'
+import deLanguages from 'cldr-localenames-full/main/de/languages.json'
+import deTerritories from 'cldr-localenames-full/main/de/territories.json'
+import deScripts from 'cldr-localenames-full/main/de/scripts.json'
+import deVariants from 'cldr-localenames-full/main/de/variants.json'
 
 // French
 import frNumbers from 'cldr-numbers-full/main/fr/numbers.json'
 import frCurrencies from 'cldr-numbers-full/main/fr/currencies.json'
 import frDates from 'cldr-dates-full/main/fr/ca-gregorian.json'
-import frLocaleNames from 'cldr-localenames-full/main/fr/languages.json'
+import frLanguages from 'cldr-localenames-full/main/fr/languages.json'
+import frTerritories from 'cldr-localenames-full/main/fr/territories.json'
+import frScripts from 'cldr-localenames-full/main/fr/scripts.json'
+import frVariants from 'cldr-localenames-full/main/fr/variants.json'
 
 // Spanish
 import esNumbers from 'cldr-numbers-full/main/es/numbers.json'
 import esCurrencies from 'cldr-numbers-full/main/es/currencies.json'
 import esDates from 'cldr-dates-full/main/es/ca-gregorian.json'
-import esLocaleNames from 'cldr-localenames-full/main/es/languages.json'
+import esLanguages from 'cldr-localenames-full/main/es/languages.json'
+import esTerritories from 'cldr-localenames-full/main/es/territories.json'
+import esScripts from 'cldr-localenames-full/main/es/scripts.json'
+import esVariants from 'cldr-localenames-full/main/es/variants.json'
 
 // Italian
 import itNumbers from 'cldr-numbers-full/main/it/numbers.json'
 import itCurrencies from 'cldr-numbers-full/main/it/currencies.json'
 import itDates from 'cldr-dates-full/main/it/ca-gregorian.json'
-import itLocaleNames from 'cldr-localenames-full/main/it/languages.json'
+import itLanguages from 'cldr-localenames-full/main/it/languages.json'
+import itTerritories from 'cldr-localenames-full/main/it/territories.json'
+import itScripts from 'cldr-localenames-full/main/it/scripts.json'
+import itVariants from 'cldr-localenames-full/main/it/variants.json'
 
 // Japanese
 import jaNumbers from 'cldr-numbers-full/main/ja/numbers.json'
 import jaCurrencies from 'cldr-numbers-full/main/ja/currencies.json'
 import jaDates from 'cldr-dates-full/main/ja/ca-gregorian.json'
-import jaLocaleNames from 'cldr-localenames-full/main/ja/languages.json'
+import jaLanguages from 'cldr-localenames-full/main/ja/languages.json'
+import jaTerritories from 'cldr-localenames-full/main/ja/territories.json'
+import jaScripts from 'cldr-localenames-full/main/ja/scripts.json'
+import jaVariants from 'cldr-localenames-full/main/ja/variants.json'
 
 // Chinese Simplified
 import zhHansNumbers from 'cldr-numbers-full/main/zh-Hans/numbers.json'
 import zhHansCurrencies from 'cldr-numbers-full/main/zh-Hans/currencies.json'
 import zhHansDates from 'cldr-dates-full/main/zh-Hans/ca-gregorian.json'
-import zhHansLocaleNames from 'cldr-localenames-full/main/zh-Hans/languages.json'
+import zhHansLanguages from 'cldr-localenames-full/main/zh-Hans/languages.json'
+import zhHansTerritories from 'cldr-localenames-full/main/zh-Hans/territories.json'
+import zhHansScripts from 'cldr-localenames-full/main/zh-Hans/scripts.json'
+import zhHansVariants from 'cldr-localenames-full/main/zh-Hans/variants.json'
 
 // Korean
 import koNumbers from 'cldr-numbers-full/main/ko/numbers.json'
 import koCurrencies from 'cldr-numbers-full/main/ko/currencies.json'
 import koDates from 'cldr-dates-full/main/ko/ca-gregorian.json'
-import koLocaleNames from 'cldr-localenames-full/main/ko/languages.json'
+import koLanguages from 'cldr-localenames-full/main/ko/languages.json'
+import koTerritories from 'cldr-localenames-full/main/ko/territories.json'
+import koScripts from 'cldr-localenames-full/main/ko/scripts.json'
+import koVariants from 'cldr-localenames-full/main/ko/variants.json'
 
 // Arabic
 import arNumbers from 'cldr-numbers-full/main/ar/numbers.json'
 import arCurrencies from 'cldr-numbers-full/main/ar/currencies.json'
 import arDates from 'cldr-dates-full/main/ar/ca-gregorian.json'
-import arLocaleNames from 'cldr-localenames-full/main/ar/languages.json'
+import arLanguages from 'cldr-localenames-full/main/ar/languages.json'
+import arTerritories from 'cldr-localenames-full/main/ar/territories.json'
+import arScripts from 'cldr-localenames-full/main/ar/scripts.json'
+import arVariants from 'cldr-localenames-full/main/ar/variants.json'
 
 // Russian
 import ruNumbers from 'cldr-numbers-full/main/ru/numbers.json'
 import ruCurrencies from 'cldr-numbers-full/main/ru/currencies.json'
 import ruDates from 'cldr-dates-full/main/ru/ca-gregorian.json'
-import ruLocaleNames from 'cldr-localenames-full/main/ru/languages.json'
+import ruLanguages from 'cldr-localenames-full/main/ru/languages.json'
+import ruTerritories from 'cldr-localenames-full/main/ru/territories.json'
+import ruScripts from 'cldr-localenames-full/main/ru/scripts.json'
+import ruVariants from 'cldr-localenames-full/main/ru/variants.json'
 
 // Portuguese
 import ptNumbers from 'cldr-numbers-full/main/pt/numbers.json'
 import ptCurrencies from 'cldr-numbers-full/main/pt/currencies.json'
 import ptDates from 'cldr-dates-full/main/pt/ca-gregorian.json'
-import ptLocaleNames from 'cldr-localenames-full/main/pt/languages.json'
+import ptLanguages from 'cldr-localenames-full/main/pt/languages.json'
+import ptTerritories from 'cldr-localenames-full/main/pt/territories.json'
+import ptScripts from 'cldr-localenames-full/main/pt/scripts.json'
+import ptVariants from 'cldr-localenames-full/main/pt/variants.json'
 
 // Hindi
 import hiNumbers from 'cldr-numbers-full/main/hi/numbers.json'
 import hiCurrencies from 'cldr-numbers-full/main/hi/currencies.json'
 import hiDates from 'cldr-dates-full/main/hi/ca-gregorian.json'
-import hiLocaleNames from 'cldr-localenames-full/main/hi/languages.json'
+import hiLanguages from 'cldr-localenames-full/main/hi/languages.json'
+import hiTerritories from 'cldr-localenames-full/main/hi/territories.json'
+import hiScripts from 'cldr-localenames-full/main/hi/scripts.json'
+import hiVariants from 'cldr-localenames-full/main/hi/variants.json'
 
 // Turkish
 import trNumbers from 'cldr-numbers-full/main/tr/numbers.json'
 import trCurrencies from 'cldr-numbers-full/main/tr/currencies.json'
 import trDates from 'cldr-dates-full/main/tr/ca-gregorian.json'
-import trLocaleNames from 'cldr-localenames-full/main/tr/languages.json'
+import trLanguages from 'cldr-localenames-full/main/tr/languages.json'
+import trTerritories from 'cldr-localenames-full/main/tr/territories.json'
+import trScripts from 'cldr-localenames-full/main/tr/scripts.json'
+import trVariants from 'cldr-localenames-full/main/tr/variants.json'
 
 // Polish
 import plNumbers from 'cldr-numbers-full/main/pl/numbers.json'
 import plCurrencies from 'cldr-numbers-full/main/pl/currencies.json'
 import plDates from 'cldr-dates-full/main/pl/ca-gregorian.json'
-import plLocaleNames from 'cldr-localenames-full/main/pl/languages.json'
+import plLanguages from 'cldr-localenames-full/main/pl/languages.json'
+import plTerritories from 'cldr-localenames-full/main/pl/territories.json'
+import plScripts from 'cldr-localenames-full/main/pl/scripts.json'
+import plVariants from 'cldr-localenames-full/main/pl/variants.json'
 
 // Core data
 import availableLocales from 'cldr-core/availableLocales.json'
+
+/**
+ * Helper function to merge separate locale names files into a single object
+ */
+function mergeLocaleNames(languages: unknown, territories: unknown, scripts: unknown, variants: unknown) {
+  const languagesData = languages as { main: { [key: string]: { localeDisplayNames: { languages: Record<string, string> } } } }
+  const territoriesData = territories as { main: { [key: string]: { localeDisplayNames: { territories: Record<string, string> } } } }
+  const scriptsData = scripts as { main: { [key: string]: { localeDisplayNames: { scripts: Record<string, string> } } } }
+  const variantsData = variants as { main: { [key: string]: { localeDisplayNames: { variants: Record<string, string> } } } }
+
+  const locale = Object.keys(languagesData.main)[0]
+
+  return {
+    main: {
+      [locale]: {
+        localeDisplayNames: {
+          languages: languagesData.main[locale].localeDisplayNames.languages,
+          territories: territoriesData.main[locale].localeDisplayNames.territories,
+          scripts: scriptsData.main[locale].localeDisplayNames.scripts,
+          variants: variantsData.main[locale].localeDisplayNames.variants,
+        }
+      }
+    }
+  }
+}
 
 export const localeData = {
   numbers: {
@@ -140,20 +207,20 @@ export const localeData = {
     pl: plDates,
   },
   localeNames: {
-    en: enLocaleNames,
-    de: deLocaleNames,
-    fr: frLocaleNames,
-    es: esLocaleNames,
-    it: itLocaleNames,
-    ja: jaLocaleNames,
-    'zh-Hans': zhHansLocaleNames,
-    ko: koLocaleNames,
-    ar: arLocaleNames,
-    ru: ruLocaleNames,
-    pt: ptLocaleNames,
-    hi: hiLocaleNames,
-    tr: trLocaleNames,
-    pl: plLocaleNames,
+    en: mergeLocaleNames(enLanguages, enTerritories, enScripts, enVariants),
+    de: mergeLocaleNames(deLanguages, deTerritories, deScripts, deVariants),
+    fr: mergeLocaleNames(frLanguages, frTerritories, frScripts, frVariants),
+    es: mergeLocaleNames(esLanguages, esTerritories, esScripts, esVariants),
+    it: mergeLocaleNames(itLanguages, itTerritories, itScripts, itVariants),
+    ja: mergeLocaleNames(jaLanguages, jaTerritories, jaScripts, jaVariants),
+    'zh-Hans': mergeLocaleNames(zhHansLanguages, zhHansTerritories, zhHansScripts, zhHansVariants),
+    ko: mergeLocaleNames(koLanguages, koTerritories, koScripts, koVariants),
+    ar: mergeLocaleNames(arLanguages, arTerritories, arScripts, arVariants),
+    ru: mergeLocaleNames(ruLanguages, ruTerritories, ruScripts, ruVariants),
+    pt: mergeLocaleNames(ptLanguages, ptTerritories, ptScripts, ptVariants),
+    hi: mergeLocaleNames(hiLanguages, hiTerritories, hiScripts, hiVariants),
+    tr: mergeLocaleNames(trLanguages, trTerritories, trScripts, trVariants),
+    pl: mergeLocaleNames(plLanguages, plTerritories, plScripts, plVariants),
   },
   core: {
     availableLocales,